Re-organize for better clarity
